wip choice-group and option add/eddit/delte
