+++ subscribie/blueprints/admin/__init__.py
@@ -19,7 +19,6 @@ from subscribie import (
     GoogleTagManagerConnectForm,
     PlansForm,
     ItemForm,
-    ChoiceGroupForm,
     jsonify,
     TawkConnectForm,
     database, User, Person, Subscription, SubscriptionNote, Company, Item,
@@ -39,9 +38,13 @@ from sqlalchemy import asc, desc
 from datetime import datetime
 from subscribie.models import ChoiceGroup, Transaction
 
+
 admin = Blueprint(
     "admin", __name__, template_folder="templates", static_folder="static"
 )
+
+from .choice_group import list_choice_groups
+from .option import list_options
 
 @admin.app_template_filter()
 def currencyFormat(value):
@@ -384,56 +387,6 @@ def delete_item(id):
     flash("Item deleted")
     return redirect(url_for('admin.list_items'))
 
-@admin.route("/add-choice-group", methods=["GET", "POST"])
-@login_required
-def add_choice_group():
-    form = ChoiceGroupForm()
-    if form.validate_on_submit():
-        choice_group = ChoiceGroup()
-        choice_group.title = request.form['title']
-        choice_group.description = request.form['description']
-        database.session.add(choice_group)
-        database.session.commit()
-        flash("Added new choice group")
-        return redirect(url_for('admin.list_choice_groups'))
-    
-    return render_template("admin/choice_group/add_choice_group.html", form=form)
-
-@admin.route("/list-choice-groups", methods=["GET", "POST"])
-@login_required
-def list_choice_groups():
-    choice_groups = ChoiceGroup.query.all()
-    return render_template("admin/choice_group/list_choice_groups.html", choice_groups=choice_groups)
-
-@admin.route("/edit-choice-group/<id>", methods=["GET", "POST"])
-@login_required
-def edit_choice_group(id):
-    choice_group = ChoiceGroup.query.get(id)
-    if request.method == 'POST':
-        choice_group.title = request.form['title']
-        choice_group.description = request.form['description']
-        database.session.commit()
-        flash("Choice group updated")
-    return render_template("admin/choice_group/edit_choice_group.html", choice_group=choice_group)
-
-@admin.route("/delete-choice-group/<id>", methods=["GET"])
-@login_required
-def delete_choice_group(id):
-    if "confirm" in request.args:
-        choice_groups = ChoiceGroup.query.all()
-
-        confirm = False
-        return render_template(
-            "admin/choice_group/list_choice_groups.html",
-            choice_groups=choice_groups,
-            choice_group=ChoiceGroup.query.get(id),
-            confirm=False,
-        )
-    choice_group = ChoiceGroup.query.get(id)
-    database.session.delete(choice_group)
-    database.session.commit()
-    flash("Choice group deleted")
-    return redirect(url_for('admin.list_choice_groups'))
 
 @admin.route("/add", methods=["GET", "POST"])
 @login_required
