+++ subscribie/blueprints/admin/templates/admin/option/add_option.html
@@ -1,0 +1,54 @@
+{% extends "admin/layout.html" %}
+{% block title %} Add Shop Admin{% endblock %}
+
+{% block body %}
+
+<h2 class="text-center text-dark mb-3">Add Option</h2>
+
+<div class="container">
+  <ul class="breadcrumb">
+    <li class="breadcrumb-item"><a href="/">Shop</a></li>
+    <li class="breadcrumb-item"><a href="{{ url_for('admin.dashboard') }}">Manage My Shop</a></li>
+    <li class="breadcrumb-item"><a href="">Options</a></li>   
+    <li class="breadcrumb-item active" aria-current="page">Add Option</li>
+  </ul>
+</div>
+
+<main>
+  <div class="section">
+    <div class="container">
+
+      <div class="row row-cols-1 row-cols-md-2">
+        <div class="col-md-7">
+          <h2>Add Option</h2>
+          <p>Options are part of choice groups.</p>
+          
+          <form action="" method="POST" action="/">
+            {{ form.csrf_token }}
+            <div class="form-group row">
+              <label class="col-sm-3 col-form-label" for="title">Title:</label>
+              <div class="col-sm-9">
+                <input class="form-control" type="title" name="title" required>
+              </div>
+            </div>
+            <div class="form-group row">
+              <label class="col-sm-3 col-form-label" for="description">Description:</label>
+              <div class="col-sm-9">
+                <textarea class="form-control" name="description" required></textarea>
+              </div>
+            </div>
+            <button type="submit" class="btn btn-primary btn-block mb-3">Save</button>
+          </form>
+        </div>
+
+        <div class="col-md-5">
+        </div>
+      </div>
+
+    </div> <!-- end container-->
+  </div> <!-- end section-->
+</main>
+
+
+
+{% endblock %}
